vr version_date const, test improvements
* organizing tests,
* cleanup temp custom classifiers
